Direct method via construction of tree shapes
Similar to mutations.py, I am trying to establish the equivalence relations within vector k. I do this here, via direct iterative construction of the tree shapes. This is the primary contribution including the outline/idea, and the code still needs to be tested. Small additions are necessary to allow it to feed into the equation-graph paradigm.